lambda/sfr-oclc-lookup/requirements.txt
boto3
lxml
marcalyx
python-lambda
python-levenshtein
pyyaml
requests
